pyproject.toml
[project]
name = "opendss_powerflow_service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.21",
    "celery>=5.4.0",
    "fastapi[standard]>=0.115.12",
    "flower>=2.0.1",
    "geojson>=3.2.0",
    "opendssdirect-py>=0.9.4",
    "pandas>=2.2.3",
    "psycopg>=3.2.6",
    "psycopg-binary>=3.2.6",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.5",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
]
